Factories generated in one file, w/ associations.
Previously, each model had its own factory file.  For my tastes, this is
too similar to fixtures.  I want all of my factories in one file
(test/factories.rb).

Previously, when the task came across 'user_id' (for example), it
would give it a value of 1 and be on its way.  I wanted it to realize
that that key was probably a foreign key, and to automatically create an
association line ('o.user { |u| u.association(:user) }').  This isn't
perfect (it's not checking AR to see if the association exists, for
instance), but it's correct more often than not.  In other words, I have
to open factories.rb a couple fewer times.

The task will now check for spec/, and put the factories.rb
file there.  If that doesn't exist, it will go in test/ (which should
always exist if spec does not -- right?).

This is one LONG freakin' commit message.